+++ core/server/common/src/main/java/alluxio/master/audit/AsyncUserAccessAuditLogWriter.java
@@ -31,8 +31,7 @@ public final class AsyncUserAccessAuditLogWriter {
   private static final String AUDIT_LOG_THREAD_NAME = "AsyncUserAccessAuditLogger";
   private static final Logger LOG =
       LoggerFactory.getLogger(AsyncUserAccessAuditLogWriter.class);
-  private static final Logger AUDIT_LOG =
-      LoggerFactory.getLogger("AUDIT_LOG");
+  private final Logger mLog;
   private volatile boolean mStopped;
   /**
    * A thread-safe linked-list-based queue with an optional capacity limit.
@@ -46,10 +45,13 @@ public final class AsyncUserAccessAuditLogWriter {
 
   /**
    * Constructs an {@link AsyncUserAccessAuditLogWriter} instance.
+   *
+   * @param loggerName the logger name
    */
-  public AsyncUserAccessAuditLogWriter() {
+  public AsyncUserAccessAuditLogWriter(String loggerName) {
     int queueCapacity = ServerConfiguration.getInt(PropertyKey.MASTER_AUDIT_LOGGING_QUEUE_CAPACITY);
     mAuditLogEntries = new LinkedBlockingQueue<>(queueCapacity);
+    mLog = LoggerFactory.getLogger(loggerName);
     LOG.info("Audit logging queue capacity is {}.", queueCapacity);
     mStopped = true;
   }
@@ -121,7 +123,7 @@ public final class AsyncUserAccessAuditLogWriter {
       while (!mStopped) {
         try {
           AuditContext headContext = mAuditLogEntries.take();
-          AUDIT_LOG.info(headContext.toString());
+          mLog.info(headContext.toString());
         } catch (InterruptedException e) {
           Thread.currentThread().interrupt();
           break;