+++ core/server/master/src/main/java/alluxio/master/metastore/rocks/RocksStore.java
@@ -101,7 +101,7 @@ public final class RocksStore implements Closeable {
     } catch (RocksDBException e) {
       throw new RuntimeException(e);
     }
-    LOG.info("Cleared store at %s", mDbPath);
+    LOG.info("Cleared store at {}", mDbPath);
   }
 
   private void resetDb() throws RocksDBException {
@@ -217,6 +217,6 @@ public final class RocksStore implements Closeable {
   @Override
   public synchronized void close() {
     stopDb();
-    LOG.info("Closed store at %s", mDbPath);
+    LOG.info("Closed store at {}", mDbPath);
   }
 }