+++ job/common/src/test/java/alluxio/job/workflow/composite/CompositeConfigTest.java
@@ -42,9 +42,8 @@ public final class CompositeConfigTest {
     ArrayList<JobConfig> jobs = new ArrayList<>();
     jobs.add(new CompositeConfig(new ArrayList<>(), true));
     jobs.add(new CompositeConfig(new ArrayList<>(), false));
-    jobs.add(new CompositeConfig(
-        Lists.newArrayList(new LoadConfig("/", 1, Collections.EMPTY_SET,
-            Collections.EMPTY_SET, Collections.EMPTY_SET, Collections.EMPTY_SET)), true));
+    jobs.add(new CompositeConfig(Lists.newArrayList(new LoadConfig("/", 1, Collections.EMPTY_SET,
+        Collections.EMPTY_SET, Collections.EMPTY_SET, Collections.EMPTY_SET, true)), true));
     jobs.add(new CompactConfig(pInfo, "/input", pInfo, "/output", 100, FileUtils.ONE_GB));
     CONFIG = new CompositeConfig(jobs, true);
   }