+++ shell/src/main/java/alluxio/cli/fs/command/DistributedLoadCommand.java
@@ -195,6 +195,14 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
           .argName("batch-size")
           .desc("Number of files per request")
           .build();
+  private static final Option PASSIVE_CACHE_OPTION =
+      Option.builder()
+          .longOpt("passive-cache")
+          .required(false)
+          .hasArg(false)
+          .desc("Use passive-cache or direct cache request,"
+              + " turn on if you want to use the old passive cache implementation")
+          .build();
 
   /**
    * Constructs a new instance to load a file or directory in Alluxio space.
@@ -216,6 +224,7 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
         .addOption(INDEX_FILE)
         .addOption(HOSTS_OPTION)
         .addOption(HOST_FILE_OPTION)
+        .addOption(PASSIVE_CACHE_OPTION)
         .addOption(BATCH_SIZE_OPTION);
   }
 
@@ -233,6 +242,7 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
         + "[--locality-file <localityFilePath>] "
         + "[--excluded-locality <locality1>,<locality2>,...,<localityN>] "
         + "[--excluded-locality-file <localityFilePath>] "
+        + "[--passive-cache] "
         + "<path>";
   }
 
@@ -251,6 +261,7 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
     int defaultBatchSize = conf.getInt(PropertyKey.JOB_REQUEST_BATCH_SIZE);
     int replication = FileSystemShellUtils.getIntArg(cl, REPLICATION_OPTION, DEFAULT_REPLICATION);
     int batchSize = FileSystemShellUtils.getIntArg(cl, BATCH_SIZE_OPTION, defaultBatchSize);
+    boolean directCache = !cl.hasOption(PASSIVE_CACHE_OPTION.getLongOpt());
     Set<String> workerSet = new HashSet<>();
     Set<String> excludedWorkerSet = new HashSet<>();
     Set<String> localityIds = new HashSet<>();
@@ -287,14 +298,14 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
     if (!cl.hasOption(INDEX_FILE.getLongOpt())) {
       AlluxioURI path = new AlluxioURI(args[0]);
       DistributedLoadUtils.distributedLoad(this, filePool, batchSize, path, replication, workerSet,
-          excludedWorkerSet, localityIds, excludedLocalityIds, true);
+          excludedWorkerSet, localityIds, excludedLocalityIds, directCache, true);
     } else {
       try (BufferedReader reader = new BufferedReader(new FileReader(args[0]))) {
         for (String filename; (filename = reader.readLine()) != null;) {
           AlluxioURI path = new AlluxioURI(filename);
 
           DistributedLoadUtils.distributedLoad(this, filePool, batchSize, path, replication,
-              workerSet, excludedWorkerSet, localityIds, excludedLocalityIds, true);
+              workerSet, excludedWorkerSet, localityIds, excludedLocalityIds, directCache, true);
         }
       }
     }